refactor(core): rename modifiers dir to decorators
