+++ src/decorators/lazy.ts
