+++ src/decorators/object.ts
