+++ src/decorators/string.ts
