+++ test/unit/decorators/func.test.ts
