+++ test/unit/decorators/lazy.test.ts
