+++ test/unit/decorators/object.test.ts
