+++ test/unit/decorators/string.test.ts
